refactor(rpc_client): move signing pool to shared _executors module (synth-007)
client_emissions imported the leading-underscore _signing_executor across a
module boundary from tx_manager, coupling the bundle-signing path to a
private implementation detail of the tx layer. Move the pool (and its size
resolution) into a small shared rpc_client/_executors module that both
tx_manager and client_emissions import from, making the cross-cutting concern
explicit without exposing it through either module's public API. Exposed as
signing_executor (the underscore module is the privacy boundary).
